requirements.txt
optuna==3.4.0
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
hydra-core==1.3.2
mlflow==2.8.1
dvc==3.30.3
pyyaml==6.0.1
psycopg2-binary==2.9.7
